cards improved

+++ src/JRoyale/Spell.java
@@ -2,7 +2,7 @@ public abstract class Spell extends Card
 {
     private final double radius;
 
-    public Spell(String id, int cost, int radius) 
+    public Spell(String id, int cost, double radius) 
     {
         super(id, cost);
         this.radius = radius;
